itoa (things missing)

+++ ft_strlen.c
@@ -6,16 +6,15 @@
 /*   By: digonza2 <digonza2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/02 13:17:00 by digonza2          #+#    #+#             */
-/*   Updated: 2025/10/03 00:23:19 by digonza2         ###   ########.fr       */
+/*   Updated: 2025/10/08 16:48:37 by digonza2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
 #include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
